chore: polish some trivals to make code clean
