+++ src/consensus/parlia/db.rs
@@ -1,8 +1,3 @@
-//! Parlia snapshot database table definitions.
-//!
-//! Stored value is the CBOR‐compressed `Snapshot` blob returned by
-//! `Compress` implementation.
-
 use reth_db::table::Table;
 
 /// Table: epoch boundary block number (u64) -> compressed snapshot bytes.
