+++ src/consensus/parlia/mod.rs
@@ -1,7 +1,6 @@
 pub mod vote;
 pub mod snapshot;
 pub mod provider;
-pub mod hertz_patch;
 pub mod constants;
 pub mod gas;
 pub mod util;
@@ -9,11 +8,9 @@ pub mod error;
 pub mod consensus;
 pub mod validation;
 pub mod db;
-
 pub use snapshot::{Snapshot, ValidatorInfo, CHECKPOINT_INTERVAL};
 pub use vote::{VoteAddress, VoteAttestation, VoteData, VoteEnvelope, VoteSignature, ValidatorsBitSet};
 pub use constants::*;
-pub use hertz_patch::{HertzPatchManager, StoragePatch};
 pub use error::ParliaConsensusError;
 pub use util::hash_with_chain_id;
 pub use provider::SnapshotProvider;
