+++ src/lib.rs
@@ -6,7 +6,6 @@ pub mod node;
 pub mod rpc;
 pub mod shared;
 pub use node::primitives::BscPrimitives;
-// Re-export the BSC-specific block types so modules can `use crate::{BscBlock, BscBlockBody, …}`
 pub use node::primitives::{BscBlock, BscBlockBody, BscBlobTransactionSidecar};
 mod system_contracts;
 pub use system_contracts::SLASH_CONTRACT;
