+++ src/shared.rs
@@ -1,8 +1,3 @@
-//! Shared global state for BSC node components
-//! 
-//! This module provides global access to the snapshot provider so that
-//! both the consensus builder and RPC modules can access the same instance.
-// TODO: refine it later.
 use crate::consensus::parlia::SnapshotProvider;
 use std::sync::{Arc, OnceLock};
 use alloy_consensus::Header;
